requirements.txt
datasets>=2.16.1
pandas>=2.2.0
beautifulsoup4>=4.12.3
html2text>=2020.1.16
tqdm>=4.66.1
ijson>=3.2.3
smart_open>=6.4.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
flask==2.0.1
flask-cors>=4.0.0
psutil==5.9.5
scikit-learn==1.2.2
tqdm==4.65.0
gunicorn==20.1.0
werkzeug==2.0.3
python-dotenv==1.0.0
requests==2.28.2
jinja2==3.0.3
itsdangerous==2.0.1
click==8.0.4
markupsafe==2.0.1
threadpoolctl==3.1.0
joblib==1.2.0
scipy==1.10.1
pillow==9.5.0
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
huggingface-hub==0.15.1
tokenizers==0.13.3
regex==2023.6.3 